LPS-120269 Reorganize packages - src
