+++ modules/apps/portal-remote/portal-remote-cors-impl/src/main/java/com/liferay/portal/remote/cors/internal/url/pattern/mapper/base/BaseURLPatternMapper.java
@@ -12,7 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.portal.remote.cors.internal.url.pattern.mapper.base;
+
+import com.liferay.portal.remote.cors.internal.url.pattern.mapper.URLPatternMapper;
 
 import java.net.URI;
 import java.net.URISyntaxException;
