+++ modules/apps/portal-remote/portal-remote-cors-impl/src/main/java/com/liferay/portal/remote/cors/internal/url/pattern/mapper/trie/DynamicSizeTrieURLPatternMapper.java
@@ -12,7 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.portal.remote.cors.internal.url.pattern.mapper.trie;
+
+import com.liferay.portal.remote.cors.internal.url.pattern.mapper.trie.base.BaseTrieURLPatternMapper;
 
 import java.util.ArrayList;
 import java.util.List;
